+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.0] - 2023-10-06
+
+### Changed
+
+- Improving the settings use case in order to be able to use app settings parameters stored in NVM
+- add a NBGL use case choice when a setting switch is toggled
+
+
 ## [2.0.0] - 2023-07-10
 
 ### Added